Allowing Cut application to work.
Note: This can only work by cutting using character position and it is working with file. This is tested with 3 possible combinations of LIST. In addition, the CutException is removed inside impl/exception as this is redundant. Furthermore, CutException have additional constructor to allow Exception to be included inside constructor to avoid PMD violations.